+++ src/strategies/horizontalLevelBounce/supportBounce.go
@@ -35,7 +35,7 @@ func SupportBounce(params strategies.Params) {
 	}
 
 	price = price + params.MarketStrategyParams.LimitAndStopOrderPriceOffset
-	if price >= float64(params.Market.GetCurrentBrokerQuote().Bid) {
+	if price >= params.Market.GetCurrentBrokerQuote().Bid {
 		params.Market.Log("Price is lower than the current ask, so we can't create the long order now. Price is -> " + utils.FloatToString(price, params.MarketData.PriceDecimals))
 		params.Market.Log("Quote is -> " + utils.GetStringRepresentation(params.Market.GetCurrentBrokerQuote()))
 		return
